requirements.txt
Django~=5.0.4
python-dotenv~=1.0.1
djangorestframework==3.15.2
psycopg2-binary==2.9.9

celery~=5.4.0
config~=0.5.1
requests~=2.31.0
django-celery-beat==2.6.0
django-cors-headers==4.4.0
django-filter==24.2
django-timezone-field==7.0
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
drf-yasg==1.21.7
eventlet==0.36.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
redis==5.0.7

pillow==10.4.0
